+++ config/routes.rb
@@ -77,7 +77,11 @@ Tahi::Application.routes.draw do
     resources :decisions, only: [:create, :update, :show] do
       put :rescind, on: :member
       put :register, on: :member
+      resources :attachments, only: [:index, :create, :update, :destroy], controller: 'decision_attachments' do
+        put :update_attachment, on: :member
+      end
     end
+    resources :decision_attachments, only: [:index, :show, :create, :update, :destroy]
     resources :discussion_topics, only: [:index, :show, :create, :update] do
       get :users, on: :member
     end
@@ -164,7 +168,7 @@ Tahi::Application.routes.draw do
     end
 
     resources :related_articles, only: [:show, :create, :update, :destroy]
-    resources :reviewer_reports, only: [:show, :create, :update, :destroy]
+    resources :reviewer_reports, only: [:show, :update]
     resources :tasks, only: [:update, :create, :show, :destroy] do
       get :nested_questions
       get :nested_question_answers